Use tools for tags, group for op.externalDocs
The group is too specific for tag for many code-generation purposes
(e.g. generating classes).  Tools (which is used as the parent in the
online docs sidebar hierarchy) seems more suitable.  Use tools for tags.

Move the group->externalDocs transform from tag to operation now that
group is not used for tags.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>