fix: transform items and properties on all schemas
If items and/or properties is present, transform them, regardless of the
declared type of the schema.

A response schema for syncPotentialChangeOrders has both items and
properties (with `type: object`).  Both should be converted to OpenAPI.
If it is an error (as I think it is) it can be cleaned up in a
transformer, rather than during conversion.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>